requirements.txt
fastapi
uvicorn
requests
pydantic
chromadb
python-dotenv
langchain
python-docx